+++ src/main/java/com/likelion/scul/auth/repository/RefreshTokenRepository.java
@@ -10,4 +10,6 @@ public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long
     Optional<RefreshToken> findByToken(String token);
 
     void deleteByUser(User user);
+
+    Optional<RefreshToken> findByUser(User user);
 }
