+++ src/main/java/com/likelion/scul/auth/service/JwtService.java
@@ -1,5 +1,8 @@
 package com.likelion.scul.auth.service;
 
+import com.likelion.scul.auth.domain.RefreshToken;
+import com.likelion.scul.auth.repository.RefreshTokenRepository;
+import com.likelion.scul.common.domain.User;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.beans.factory.annotation.Value;
@@ -7,12 +10,19 @@ import org.springframework.stereotype.Service;
 
 import java.security.Key;
 import java.util.Date;
+import java.util.Optional;
 
 @Service
 public class JwtService {
 
     @Value("${jwt.secret}")
     private String secretKey;
+
+    private RefreshTokenRepository refreshTokenRepository;
+
+    public JwtService(RefreshTokenRepository refreshTokenRepository) {
+        this.refreshTokenRepository = refreshTokenRepository;
+    }
 
     public static final long ACCESS_TOKEN_VALIDITY = 1000 * 60 * 60; // 1시간
     public static final long REFRESH_TOKEN_VALIDITY = 1000 * 60 * 60 * 24 * 7; // 7일
@@ -41,6 +51,15 @@ public class JwtService {
         return "Bearer " + token;
     }
 
+    public RefreshToken createAndSaveRefreshToken(User user, String token) {
+        RefreshToken refreshToken = new RefreshToken();
+        refreshToken.setUser(user);
+        refreshToken.setToken(token);
+        refreshToken.setExpiryDate(new Date(System.currentTimeMillis() + JwtService.REFRESH_TOKEN_VALIDITY));
+
+        return refreshTokenRepository.save(refreshToken);
+    }
+
     public boolean validateToken(String token) {
         try {
             token = removeBearerPrefix(token);
@@ -57,16 +76,29 @@ public class JwtService {
         }
     }
 
-    public Claims getClaimsFromToken(String token) {
-        token = removeBearerPrefix(token);
-        Jws<Claims> claimsJws = Jwts.parserBuilder().setSigningKey(getSigningKey()).build().parseClaimsJws(token);
-        return claimsJws.getBody();
-    }
-
     private String removeBearerPrefix(String token) {
         if (token.startsWith("Bearer ")) {
             return token.substring(7);
         }
         return token;
+    }
+
+    public Claims getClaimsFromToken(String token) {
+        Jws<Claims> claimsJws = Jwts.parserBuilder().setSigningKey(getSigningKey()).build().parseClaimsJws(token);
+        return claimsJws.getBody();
+    }
+
+    public void deleteRefreshToken(User user) {
+        refreshTokenRepository.deleteByUser(user);
+    }
+
+    public String findByUser(User user) {
+        RefreshToken token = refreshTokenRepository.findByUser(user)
+                .orElseThrow(() -> new IllegalStateException("user의 refreshToken이 존재하지 않습니다."));
+        return token.getToken();
+    }
+
+    public Optional<RefreshToken> findByToken(String token) {
+        return refreshTokenRepository.findByToken(token);
     }
 }
